Move page-view-element to components/core
