+++ src/components/core/auth-controller.ts
@@ -3,7 +3,7 @@
 import { ReactiveController, ReactiveControllerHost } from 'lit';
 import { RootState } from '../../app/store.js';
 import { selectCurrentUserId } from '../../slices/auth/index.js';
-import { AuthorizedView } from '../page-view-element.js';
+import { AuthorizedView } from './page-view-element.js';
 import { StateSubscribedController } from './connect-mixin.js';
 
 type AuthorizedViewHost = ReactiveControllerHost & AuthorizedView;
