+++ src/components/lineup-view-game-detail.ts
@@ -9,10 +9,10 @@ import { GameDetail, GameStatus } from '../models/game.js';
 import { getGame, getGameSliceConfigurator, selectGameById } from '../slices/game/index.js';
 import { SignedInAuthController } from './core/auth-controller.js';
 import { ConnectStoreMixin } from './core/connect-mixin.js';
+import { AuthorizedViewElement } from './core/page-view-element.js';
 import './lineup-game-complete.js';
 import './lineup-game-live.js';
 import './lineup-game-setup.js';
-import { AuthorizedViewElement } from './page-view-element.js';
 import { SharedStyles } from './shared-styles.js';
 
 @customElement('lineup-view-game-detail')
